renamed `addResponsiveHandlers` to `addEventHandlers`
