+++ tests/test_addEventHandlers.js
@@ -1,21 +1,21 @@
 // local imports
-import addResponsiveHandlers from 'util/addResponsiveHandlers'
+import addEventHandlers from 'util/addEventHandlers'
 
 
-describe('addResponsiveHandlers', function () {
+describe('addEventHandlers', function () {
     it('returns the store it was given', function () {
         // fake store
         const store = {dispatch: () => ({})}
 
         // should return same value it is given
-        expect(addResponsiveHandlers(store)).to.equal(store)
+        expect(addEventHandlers(store)).to.equal(store)
     })
 
 
     it('calls the dispatch property of the store arg', function () {
         const dispatchSpy = sinon.spy()
 
-        addResponsiveHandlers({dispatch: dispatchSpy})
+        addEventHandlers({dispatch: dispatchSpy})
 
         // should have triggered our dispatch spy exactly once
         expect(dispatchSpy).to.have.been.calledOnce
